backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.1
django-cors-headers==4.7.0
openai==1.102.0
python-dotenv==1.1.1
Pillow==10.4.0
requests==2.31.0
beautifulsoup4==4.12.2
